+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 
 - **Bug Fix:** Multisite installations inside a sub-directory were broken by Apache Optimizations via `.htaccess` in some scenarios. Fixed in this release. See also [issue #798](https://github.com/websharks/comet-cache/issues/798) if you'd like additional details.
 
+- **Minor UI Enhancement:** This release improves the toggle link that allows you to see additional details whenever Comet Cache automatically clears more than one facet of the cache. See also [issue #837](https://github.com/websharks/comet-cache/issues/837) if you'd like additional details.
+
 = v160917 =
 
 - **New Feature** (Lite): The Clear Cache button is now available in the Admin Toolbar for the Lite version of Comet Cache.